+++ src/main/java/skd/chalba/common/TaskParams.java
@@ -10,4 +10,5 @@ public class TaskParams {
     public long _ThreadSpwanDelay;
     public String _TaskName;
 
+
 }