+++ src/main/java/skd/chalba/interfaces/TimeoutCallback.java
@@ -1,0 +1,10 @@
+package skd.chalba.interfaces;
+
+/**
+ * @author sapan.dang
+ */
+public interface TimeoutCallback {
+
+    public void timeout();
+
+}