+++ src/main/java/skd/chalba/server/ReporterSocket.java
@@ -1,0 +1,10 @@
+package skd.chalba.server;
+
+import org.eclipse.jetty.websocket.api.annotations.WebSocket;
+
+/**
+ * @author sapan.dang
+ */
+@WebSocket
+public class ReporterSocket {
+}