+++ src/test/java/skd/test/TemplateScript.java
@@ -3,7 +3,6 @@ package skd.test;
 
 import skd.chalba.common.*;
 import skd.chalba.requests.*;
-import skd.chalba.runner.*;
 import skd.chalba.interfaces.*;
 
 /**